Extract log to types module
This also removes the `CompilesTo` constraint which was an artifact of
how I arrived at this lovely little thing.
