+++ src/game/log.rs
@@ -1,54 +1,8 @@
-
-use std::{fmt::{self, Debug, Formatter}, rc::Rc};
-use super::{action::chess::ChessAction, compiles_to::CompilesTo};
-
-///////////////////////////////////////////////////////////////////////////
-/////////////////////////// TRANSACTION ///////////////////////////////////
-///////////////////////////////////////////////////////////////////////////
-
-#[derive(Clone)]
-pub struct Transaction<T> {
-    content: Vec<T>,
-    finished: bool
-}
-
-impl<T: Clone> Transaction<T> {
-    pub fn begin() -> Self {
-        Self {
-            content: vec![],
-            finished: false
-        }
-    }
-
-    pub fn is_finished(&self) -> bool {
-        self.finished
-    }
-
-    pub fn record(&mut self, action: T) {
-        self.content.push(action);
-    }
-
-    pub fn commit(&mut self) -> Vec<T> {
-        self.finished = true;
-        self.content.clone()
-    }
-
-    #[cfg(test)]
-    pub fn content(&self) -> Vec<T> {
-        self.content.clone()
-    }
-}
-
-impl<T, R> Default for Log<T, R> where T: CompilesTo<R> + Clone, R : Clone {
-    fn default() -> Self {
-        Self::start()
-    }
-}
-
 ///////////////////////////////////////////////////////////////////////////
 ///////////////////////////   CACHE   /////////////////////////////////////
 ///////////////////////////////////////////////////////////////////////////
 
+/*
 #[derive(Clone)]
 pub struct Cache<T, R> where T: CompilesTo<R> + Clone, R : Clone {
     pub entry: T,
@@ -104,314 +58,6 @@ impl<T, R> Cache<T, R> where T: CompilesTo<R> + Clone, R : Clone {
     }
 }
 
-///////////////////////////////////////////////////////////////////////////
-///////////////////////////   LOG     /////////////////////////////////////
-///////////////////////////////////////////////////////////////////////////
-
-pub struct Log<T, R> where T: CompilesTo<R> + Clone, R : Clone {
-    log: Vec<Cache<T, R>>,
-    // TODO: Extract this
-    current_txn: Transaction<T>,
-    stack: Vec<Transaction<T>>,
-
-    write_head: usize
-}
-
-impl<T, R> Log<T, R> where T: CompilesTo<R> + Clone, R : Clone {
-    pub fn start() -> Self {
-        Self {
-            log: vec![],
-            current_txn: Transaction::begin(),
-            stack: vec![],
-            write_head: 0
-        }
-    }
-
-    pub fn seek(&mut self, position: usize) -> Option<&mut Cache<T, R>> {
-        if position >= self.log.len() {
-            None
-        } else {
-            self.write_head = position;
-            self.read()
-        }
-    }
-
-    pub fn begin(&mut self) -> &mut Self {
-        self.stack.push(self.current_txn.clone());
-        self.current_txn = Transaction::begin();
-        self
-    }
-
-    pub fn record(&mut self, action: T) -> &mut Self {
-        self.current_txn.record(action);
-        self
-    }
-
-    pub fn commit(&mut self) -> &mut Self {
-        if self.stack.is_empty() {
-            // then we aren't nested.
-            let actions = self.current_txn.commit();
-            for action in actions.into_iter() {
-                let cache = Cache::from(action);
-                self.write(cache);
-            }
-
-            self.current_txn = Transaction::begin();
-        } else {
-            // then we are nested
-            let actions = self.current_txn.commit();
-
-            self.current_txn = self.stack.pop().unwrap();
-
-            for action in actions.into_iter() {
-                self.record(action);
-            }
-        }
-        self
-    }
-
-    fn write(&mut self, cache: Cache<T, R>) {
-        self.log.insert(self.write_head, cache);
-        self.write_head += 1;
-    }
-
-    fn read(&mut self) -> Option<&mut Cache<T, R>> {
-        self.log.get_mut(self.write_head)
-    }
-
-    pub fn get(&mut self, position: usize) -> Option<&mut Cache<T, R>> {
-        self.log.get_mut(position)
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.log.is_empty()
-    }
-
-    pub fn len(&self) -> usize {
-        self.log.len()
-    }
-
-    /// `` TODO: Get this working.
-    /// # use hazel::game::log::Log;
-    /// # use hazel::game::compiles_to::CompilesTo;
-    ///
-    /// let mut log = Log::default();
-    /// log.record(1).record(2).commit();
-    /// log.record(3).record(4).commit();
-    ///
-    /// log.cursor(|cursor| {
-    ///     let current = cursor
-    ///         .read()     // read the value under the cursor
-    ///         .unwrap()   // unwrap it as it may not exist if the cursor is past the end of the log
-    ///         .get(&());  // the value is in a cache that expects a compile context, in this
-    ///                     //   trivial example we don't need one, so we pass unit.
-    ///     assert_eq!(current, Some(1));
-    /// });
-    /// // The cursor is single-use.
-    /// ``
-    pub fn cursor(&mut self, block: impl Fn(&mut Cursor<T, R>)) {
-        let mut cursor = Cursor::new(self);
-        block(&mut cursor);
-    }
-
-    #[cfg(test)]
-    fn log(&self) -> Vec<T> {
-        let copy = self.log.clone();
-
-        copy.into_iter().map(|cache| cache.entry).collect::<Vec<T>>()
-    }
-
-    #[cfg(test)]
-    fn txn_state(&self) -> Vec<T> {
-        self.current_txn.content.clone()
-    }
-}
-
-impl<T, R> IntoIterator for Log<T, R> where T: CompilesTo<R> + Clone, R : Clone {
-    type Item = Cache<T, R>;
-    type IntoIter = std::vec::IntoIter<Self::Item>;
-
-    fn into_iter(self) -> Self::IntoIter {
-        self.log.into_iter()
-    }
-}
-
-///////////////////////////////////////////////////////////////////////////
-///////////////////////////   CURSOR  /////////////////////////////////////
-///////////////////////////////////////////////////////////////////////////
-
-pub struct Cursor<'a, T, R> where T: CompilesTo<R> + Clone, R : Clone {
-    log: &'a mut Log<T, R>,
-    position: usize
-}
-
-impl<'a, T, R> Cursor<'a, T, R>  where T: CompilesTo<R> + Clone, R : Clone {
-    pub fn new(log: &'a mut Log<T, R>) -> Self {
-        Self {
-            log,
-            position: 0
-        }
-    }
-
-    pub fn seek(&mut self, position: usize) -> Option<&mut Cache<T, R>> {
-        self.position = position;
-        self.read()
-    }
-
-    // FIXME: Should this be conditional?
-    pub fn jump(&mut self, offset: isize) -> Option<&mut Cache<T, R>> {
-        let new_position = self.position as isize + offset;
-
-        if new_position < 0 {
-            self.position = 0;
-            None
-        } else {
-            if new_position as usize >= self.log.log.len() {
-                self.position = self.log.len();
-            } else {
-                self.position = new_position as usize;
-            }
-            self.read()
-        }
-    }
-
-    /// Now listen for a second, I know this looks bad.
-    ///
-    /// I'm not implementing `Iterator` because of weird lifetime stuff that happens that I'm too
-    /// scared to try to fix.
-    ///
-    /// It looks bad because it is bad. Don't be like me, be brave.
-    #[allow(clippy::should_implement_trait)]
-    pub fn next(&mut self) -> Option<&mut Cache<T, R>> {
-        if self.position == self.log.len() {
-            None
-        } else {
-            self.position += 1;
-            self.read()
-        }
-    }
-
-    pub fn prev(&mut self) -> Option<&mut Cache<T, R>> {
-        if self.position == 0 {
-            None
-        } else {
-            self.position -= 1;
-            self.read()
-        }
-    }
-
-    pub fn read(&mut self) -> Option<&mut Cache<T, R>> {
-        self.log.get(self.position)
-    }
-}
-
-///////////////////////////////////////////////////////////////////////////
-///////////////////////////////////////////////////////////////////////////
-///////////////////////////////////////////////////////////////////////////
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// A trivial implementation to test with.
-    impl CompilesTo<i32> for i32 {
-        fn compile(&self, _: &()) -> i32 {
-            *self
-        }
-    }
-
-    mod cursor {
-        use super::*;
-
-        #[test]
-        fn cursor_seeks() {
-            let mut log = Log::default();
-
-            log.record(1).record(2).commit();
-
-            log.cursor(|cursor| {
-                assert_eq!(cursor.seek(0).unwrap().get(&()), Some(1));
-                assert_eq!(cursor.seek(1).unwrap().get(&()), Some(2));
-                assert_eq!(cursor.seek(2), None);
-                assert_eq!(cursor.seek(3), None);
-            });
-
-            log.record(3).record(4).commit();
-
-            log.cursor(|cursor| {
-                assert_eq!(cursor.seek(0).unwrap().get(&()), Some(1));
-                assert_eq!(cursor.seek(1).unwrap().get(&()), Some(2));
-                assert_eq!(cursor.seek(3).unwrap().get(&()), Some(4));
-                assert_eq!(cursor.seek(2).unwrap().get(&()), Some(3));
-                assert_eq!(cursor.seek(4), None);
-            });
-        }
-
-        #[test]
-        fn cursor_prev_and_next() {
-            let mut log = Log::default();
-
-            log.record(1).record(2).commit();
-
-            log.cursor(|cursor| {
-                assert_eq!(cursor.read().unwrap().get(&()), Some(1));
-                assert_eq!(cursor.next().unwrap().get(&()), Some(2));
-                assert_eq!(cursor.next(), None);
-                assert_eq!(cursor.prev().unwrap().get(&()), Some(2));
-                assert_eq!(cursor.prev().unwrap().get(&()), Some(1));
-                assert_eq!(cursor.prev(), None);
-            });
-
-            log.record(3).record(4).commit();
-
-            log.cursor(|cursor| {
-                assert_eq!(cursor.read().unwrap().get(&()), Some(1));
-                assert_eq!(cursor.next().unwrap().get(&()), Some(2));
-                assert_eq!(cursor.next().unwrap().get(&()), Some(3));
-                assert_eq!(cursor.next().unwrap().get(&()), Some(4));
-                assert_eq!(cursor.next(), None);
-                assert_eq!(cursor.prev().unwrap().get(&()), Some(4));
-                assert_eq!(cursor.prev().unwrap().get(&()), Some(3));
-                assert_eq!(cursor.prev().unwrap().get(&()), Some(2));
-                assert_eq!(cursor.prev().unwrap().get(&()), Some(1));
-                assert_eq!(cursor.prev(), None);
-            });
-        }
-
-
-        #[test]
-        fn cursor_jumps() {
-            let mut log = Log::default();
-
-            log.record(1).record(2).commit();
-
-            log.cursor(|cursor| {
-                assert_eq!(cursor.seek(0).unwrap().get(&()), Some(1));
-                assert_eq!(cursor.jump(1).unwrap().get(&()), Some(2));
-                assert_eq!(cursor.jump(1), None);
-                assert_eq!(cursor.jump(-1).unwrap().get(&()), Some(2));
-                assert_eq!(cursor.jump(-1).unwrap().get(&()), Some(1));
-                assert_eq!(cursor.jump(-1), None);
-            });
-
-            log.record(3).record(4).commit();
-
-            log.cursor(|cursor| {
-                assert_eq!(cursor.seek(0).unwrap().get(&()), Some(1));
-                assert_eq!(cursor.jump(1).unwrap().get(&()), Some(2));
-                assert_eq!(cursor.jump(1).unwrap().get(&()), Some(3));
-                assert_eq!(cursor.jump(1).unwrap().get(&()), Some(4));
-                assert_eq!(cursor.jump(1), None);
-                assert_eq!(cursor.jump(-1).unwrap().get(&()), Some(4));
-                assert_eq!(cursor.jump(-1).unwrap().get(&()), Some(3));
-                assert_eq!(cursor.jump(-1).unwrap().get(&()), Some(2));
-                assert_eq!(cursor.jump(-1).unwrap().get(&()), Some(1));
-                assert_eq!(cursor.jump(-1), None);
-            });
-        }
-
-
-    }
 
     mod cache {
         use super::*;
@@ -425,60 +71,4 @@ mod tests {
             assert_eq!(cache.get_store(), Some(1));
         }
     }
-
-    mod txn {
-        use super::*;
-
-        #[test]
-        fn commit_records_changes() {
-            let mut log = Log::default();
-            log.record(1).record(2);
-            assert_eq!(log.log(), vec![]);
-            log.commit();
-            assert_eq!(log.log(), vec![1, 2]);
-        }
-
-        #[test]
-        fn multiple_txns() {
-            let mut log = Log::default();
-            log.record(1).record(2);
-            log.commit();
-            log.record(3).record(4);
-            log.commit();
-            assert_eq!(log.log(), vec![1, 2, 3, 4]);
-        }
-
-        #[test]
-        fn nested_txns() {
-            let mut log = Log::default();
-
-            assert_eq!(log.log(), vec![]);
-
-            log.record(1).record(2);
-            log.commit();
-
-            assert_eq!(log.log(), vec![1, 2]);
-
-            log.record(3);
-
-            assert_eq!(log.txn_state(), vec![3]);
-
-            log.begin()
-                .record(4);
-
-            assert_eq!(log.txn_state(), vec![4]);
-
-            assert_eq!(log.log(), vec![1, 2]);
-
-            log.commit();
-
-            assert_eq!(log.txn_state(), vec![3, 4]);
-
-            assert_eq!(log.log(), vec![1, 2]);
-
-            log.commit();
-
-            assert_eq!(log.log(), vec![1, 2, 3, 4]);
-        }
-    }
-}
+*/
